Change bitfields to unsigned (int) (#100)
Use unsigned int for bitfields for C99 compliance

+++ nsdl-c/sn_nsdl_lib.h
@@ -146,7 +146,7 @@ typedef struct sn_nsdl_static_resource_parameters_ {
     uint16_t    resourcelen;                /**< 0 if dynamic resource, resource information in static resource */
     bool        external_memory_block:1;    /**< 0 means block messages are handled inside this library,
                                                  otherwise block messages are passed to application */
-    uint8_t     mode:2;                     /**< STATIC etc.. */
+    unsigned    mode:2;                     /**< STATIC etc.. */
     bool        free_on_delete:1;           /**< 1 if struct is dynamic allocted --> to be freed */
 } sn_nsdl_static_resource_parameters_s;
 
@@ -165,10 +165,10 @@ typedef struct sn_nsdl_resource_parameters_ {
 #endif
     ns_list_link_t                              link;
     uint16_t                                    coap_content_type;  /**< CoAP content type */
-    uint8_t                                     access:4;           /**< Allowed operation mode, GET, PUT, etc,
+    unsigned                                    access:4;           /**< Allowed operation mode, GET, PUT, etc,
                                                                          TODO! This should be in static struct but current
                                                                          mbed-client implementation requires this to be changed at runtime */
-    uint8_t                                     registered:2;       /**< Is resource registered or not */
+    unsigned                                    registered:2;       /**< Is resource registered or not */
     bool                                        publish_uri:1;      /**< 1 if resource to be published to server */
     bool                                        free_on_delete:1;   /**< 1 if struct is dynamic allocted --> to be freed */
     bool                                        observable:1;       /**< Is resource observable or not */
